Fix throwing an exception if path is invalid

+++ src/cpp/binout.cpp
@@ -81,7 +81,7 @@ Binout::get_children(const std::string &path) const noexcept {
 size_t Binout::get_num_timesteps(const std::string &path) const {
   const size_t num_timesteps{binout_get_num_timesteps(&m_handle, path.c_str())};
 
-  if (num_timesteps == ~0) {
+  if (num_timesteps == (size_t)~0) {
     throw Exception(String(
         const_cast<char *>("The path does not exist or has files as children"),
         false));
